requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# NOTE: Removed unused dependencies (playwright, beautifulsoup4, lxml, nest-asyncio)
# Old browser-based scraper (src/scraper/wegmans_scraper.py) has been replaced
# with direct Algolia API calls (src/scraper/algolia_direct.py)

# NEW: Web framework & database
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
psycopg2-binary>=2.9.9  # PostgreSQL driver
pydantic>=2.5.0

# Authentication
supabase>=2.0.0  # Supabase Python client
pyjwt>=2.8.0  # JWT token handling
email-validator>=2.1.0  # Email validation for Pydantic

# Rate limiting
slowapi>=0.1.9  # Rate limiting middleware

# Testing (dev only)
pytest>=7.4.0  # Testing framework
pytest-cov>=4.1.0  # Coverage reporting
pytest-timeout>=2.2.0  # Test timeouts
httpx>=0.25.0  # Required for FastAPI TestClient
